+++ crates/ruma-common/src/events/room/aliases.rs
@@ -2,7 +2,6 @@
 
 use ruma_macros::EventContent;
 use serde::{Deserialize, Serialize};
-use serde_json::value::RawValue as RawJsonValue;
 
 use crate::{
     events::{
@@ -82,16 +81,6 @@ impl EventContent for RedactedRoomAliasesEventContent {
 
     fn event_type(&self) -> StateEventType {
         StateEventType::RoomAliases
-    }
-
-    fn from_parts(event_type: &str, content: &RawJsonValue) -> serde_json::Result<Self> {
-        if event_type != "m.room.aliases" {
-            return Err(::serde::de::Error::custom(format!(
-                "expected event type `m.room.aliases`, found `{event_type}`",
-            )));
-        }
-
-        serde_json::from_str(content.get())
     }
 }